firecrawl_qa_module/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0

# Logging and utilities
coloredlogs>=15.0.1

# Optional: For advanced text processing
nltk>=3.8.1
spacy>=3.7.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
